requirements-dev.txt
# Dev-only dependencies to make test collection pass locally
pyyaml
spherov2
bleak
paho-mqtt
pytest
